+++ src/react-components/src/components/02-molecules/menus/Breadcrumbs/Breadcrumbs.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 import bem from "../../../../utils/bem";
+import Icon from "../../../01-atoms/Images/Icons/Icon";
 
-type breadcrumb = { url: string, text: string };
+type breadcrumb = {url: string, text: string};
 
 function isTextBreadcrumb(obj: breadcrumb | JSX.Element): obj is breadcrumb {
   return typeof((obj as breadcrumb).url) === "string" ;
@@ -19,7 +20,6 @@ export default class Breadcrumb extends React.Component<BreadcrumbProps, {}> {
   }
 
   render(): JSX.Element {
-
     const { breadcrumbs } = this.props;
 
     let breadcrumbs__base_class = "breadcrumbs";
@@ -28,15 +28,13 @@ export default class Breadcrumb extends React.Component<BreadcrumbProps, {}> {
 
     breadcrumbs.forEach((item: breadcrumb|JSX.Element, index: number) => {
       let last = index === breadcrumbs.length - 1;
-      if (last) {
-        // TODO: SFR-716 Incorporate Icon
-      }
       let linkComponent: string | JSX.Element;
+
       if (isTextBreadcrumb(item)) {
         linkComponent = item.url ? <a href={item.url} className={bem("link", [], breadcrumbs__base_class)}>{item.text}</a> : item.text;
       } else {
         let props = {
-          className: bem("link", [], breadcrumbs__base_class),
+          className: bem("link", [], breadcrumbs__base_class)
         };
         linkComponent = React.createElement(
           item.type,
@@ -44,7 +42,11 @@ export default class Breadcrumb extends React.Component<BreadcrumbProps, {}> {
           item.props.children
         );
       }
-      breadcrumbItems.push(<li key={`breadcrumb-${index}`} className={bem("item", [], breadcrumbs__base_class)} >
+      breadcrumbItems.push(<li key={`${breadcrumbs__base_class}-${index}`} className={bem("item", [], breadcrumbs__base_class)} >
+      {last && <Icon
+          name="arrow-xsmall"
+          blockName={breadcrumbs__base_class}
+        modifiers={["small"]} decorative={true} />}
       {linkComponent}
     </ li >);
     });