+++ src/react-components/CHANGELOG.md
@@ -6,6 +6,14 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 
 ========
 ## Prerelease
+
+## [ 0.0.4 ]
+### Changed 
+- `Breadcrumb` now shows icon in mobile view
+
+### Added
+- `Icon` component that uses Icons.
+
 ## [0.0.3] - 2019-11-15
 ### Changed
 - Changed `Breadcrumb` to `Breadcrumbs` to match Twig